requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.7
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-tinymce4-lite==1.7.4
gunicorn==19.9.0
idna==2.8
jsmin==2.2.2
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
urllib3>=1.24.2
whitenoise==4.1.2
